restructured project folders
