+++ src/SFA.DAS.Approvals.UITests/Project/Helpers/DataHelpers/ApprovalsEnums.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SFA.DAS.Approvals.UITests.Project.Helpers.DataHelpers
+{
+    internal enum ApprenticeshipStatus
+    {
+        Live,
+        WaitingToStart,
+        Paused,
+        Stopped,
+        Completed
+    }
+
+    internal enum FundingType
+    {
+        DirectTransferFundsFromConnection,
+        ReservedFunds,
+        ReserveNewFunds,
+        CurrentLevyFunds,
+        TransferFunds
+    }
+
+    internal enum EmployerType
+    {
+        Levy,
+        NonLevy,
+        NonLevyUserAtMaxReservationLimit
+    }
+
+}
