+++ src/SFA.DAS.Approvals.UITests/Project/Pages/ApprovalsBasePage.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SFA.DAS.Approvals.UITests.Project.Pages
+{
+    public abstract class ApprovalsBasePage(ScenarioContext context) : BasePage(context)
+    {
+        //protected readonly ApprovalsDataHelpers managingStandardsDataHelpers = context.Get<ManagingStandardsDataHelpers>();
+    }
+
+}
