+++ src/SFA.DAS.Approvals.UITests/Project/Pages/Provider/AddApprenticeDetails_SelectJourneyPage.cs
@@ -1,0 +1,26 @@
+﻿using SFA.DAS.ProviderLogin.Service.Project.Pages;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SFA.DAS.Approvals.UITests.Project.Pages.Provider
+{
+    internal class AddApprenticeDetails_SelectJourneyPage(ScenarioContext context) : ApprovalsProviderBasePage(context)
+    {
+        public override async Task VerifyPage()
+        {
+            await Assertions.Expect(page.Locator("h1")).ToContainTextAsync("Add apprentice details");
+        }
+
+        public async Task<ChooseAnEmployerPage> SelectOptionCreateANewCohort()
+        {
+            await page.GetByRole(AriaRole.Radio, new() { Name = "Create a new cohort" }).CheckAsync();
+
+            await page.GetByRole(AriaRole.Button, new() { Name = "Continue" }).ClickAsync();
+
+            return await VerifyPageAsync(() => new ChooseAnEmployerPage(context));
+        }
+    }
+}
