+++ src/SFA.DAS.Approvals.UITests/Project/Pages/Provider/ApprovalsProviderBasePage.cs
@@ -1,0 +1,14 @@
+﻿using SFA.DAS.Approvals.UITests.Project.Pages;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SFA.DAS.Approvals.UITests.Project.Pages.Provider
+{
+    public abstract class ApprovalsProviderBasePage(ScenarioContext context) : ApprovalsBasePage(context)
+    {
+
+    }
+}
