+++ src/SFA.DAS.Approvals.UITests/Project/Steps/AP_Steps.cs
@@ -1,0 +1,12 @@
+﻿using SFA.DAS.ProviderLogin.Service.Project.Helpers;
+
+namespace SFA.DAS.Approvals.UITests.Project.Steps;
+
+[Binding]
+public class AP_Steps(ScenarioContext context)
+{
+    /*
+    [Given(@"the provider logs into portal")]
+    public async Task GivenTheProviderLogsIntoPortal() => await new ProviderHomePageStepsHelper(context).GoToProviderHomePage(false);
+    */
+}
