+++ src/SFA.DAS.Approvals.UITests/Project/Steps/EmployerSteps.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SFA.DAS.Approvals.UITests.Project.Steps
+{
+    [Binding]
+    public class EmployerSteps(ScenarioContext context)
+    {
+    }
+}
