--- src/SFA.DAS.Approvals.UITests/Project/Tests/Features/01_PreApprovals/01_LevyFunds/b_ExistingEmployer/AP_E2E_EUA_01_ExistingUserAccount.feature
@@ -1,19 +1,0 @@
-﻿@approvals
-Feature: AP_E2E_EUA_01_ExistingUserAccount
-
-A short summary of the feature
-
-@regression
-@e2escenarios
-@selectstandardwithmultipleoptions
-Scenario: AP_E2E_EUA_01 Provider creates cohort from ILR data Employer approves it
- 
-	Given Provider successfully submits 1 ILR record containing a learner record for a "Levy" Employer
-	And SLD push its data into AS
-	When Provider logs into Provider-Portal
-	And creates an apprentice request (cohort) by selecting same apprentices
-	#Then Provider can send it to the Employer for approval
-	#And apprentice request (cohort) is available under 'Apprentice requests > With employers'
-	#When Employer approves the cohort
-	#Then apprentice request (cohort) is no longer available under any tab in 'Apprentice requests' section
-	#And apprentice record is available on Apprenticeships endpoint for SLD
