--- src/SFA.DAS.Approvals.UITests/Project/Tests/Pages/ApprovalsBasePage.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SFA.DAS.Approvals.UITests.Project.Tests.Pages
-{
-    public abstract class ApprovalsBasePage(ScenarioContext context) : BasePage(context)
-    {
-        //protected readonly ApprovalsDataHelpers managingStandardsDataHelpers = context.Get<ManagingStandardsDataHelpers>();
-    }
-
-}
