--- src/SFA.DAS.Approvals.UITests/Project/Tests/Pages/Provider/ApprovalsProviderBasePage.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SFA.DAS.Approvals.UITests.Project.Tests.Pages.Provider
-{
-    public abstract class ApprovalsProviderBasePage(ScenarioContext context) : ApprovalsBasePage(context)          
-    {
-
-    }
-}
