--- src/SFA.DAS.Approvals.UITests/Project/Tests/Steps/AP_Steps.cs
@@ -1,12 +1,0 @@
-﻿using SFA.DAS.ProviderLogin.Service.Project.Helpers;
-
-namespace SFA.DAS.Approvals.UITests.Project.Tests.Steps;
-
-[Binding]
-public class AP_Steps(ScenarioContext context)
-{
-    /*
-    [Given(@"the provider logs into portal")]
-    public async Task GivenTheProviderLogsIntoPortal() => await new ProviderHomePageStepsHelper(context).GoToProviderHomePage(false);
-    */
-}
