--- src/SFA.DAS.Approvals.UITests/Project/Tests/Steps/EmployerSteps.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SFA.DAS.Approvals.UITests.Project.Tests.Steps
-{
-    [Binding]
-    public class EmployerSteps(ScenarioContext context)
-    {
-    }
-}
